requirements.txt
pyttsx3>=2.90
transformers>=4.35.2
Pillow>=10.1.0